tech(pat-recurrence): lazify pattern, split registration from implementation
pat-recurrence was the single largest pattern in the eager patterns chunk
(~52KB src, incl. three statically-imported XML templates and the Modal
dependency), loaded on every page though the widget only appears on Event
add/edit forms.

Split it into a thin registration module (trigger, parser, and an init()
that dynamically imports the implementation on first match) plus
recurrence--implementation.js holding the widget, the XML templates and the
Modal import. Follows the existing tinymce--implementation lazy-pattern
convention. Public options and markup contract are unchanged.

The implementation now loads only on pages with a .pat-recurrence element,
in its own chunk.
